+++ rocketgram/api/callback_query.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-import warnings
 from dataclasses import dataclass
 from typing import Optional
 
@@ -36,9 +35,3 @@ class CallbackQuery:
 
         return cls(data['id'], User.parse(data['from']), Message.parse(data.get('message')),
                    data.get('inline_message_id'), data['chat_instance'], data.get('data'), data.get('game_short_name'))
-
-    @property
-    def query_id(self) -> str:
-        warnings.warn("This field is deprecated. Use `id` instead.", DeprecationWarning)
-
-        return self.id