+++ rocketgram/api/message_entity.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-import warnings
 from dataclasses import dataclass
 from typing import Optional
 
@@ -37,9 +36,3 @@ class MessageEntity:
 
         return cls(entity_type, data['offset'], data['length'], data.get('url'),
                    User.parse(data.get('user')), data.get('language'))
-
-    @property
-    def entity_type(self) -> EntityType:
-        warnings.warn("This field is deprecated. Use `type` instead.", DeprecationWarning)
-
-        return self.type