+++ rocketgram/api/pre_checkout_query.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-import warnings
 from dataclasses import dataclass
 from typing import Optional
 
@@ -37,9 +36,3 @@ class PreCheckoutQuery:
         return cls(data['id'], User.parse(data['from']), data['currency'], data['total_amount'],
                    data['invoice_payload'], data.get('shipping_option_id'),
                    OrderInfo.parse(data.get('order_info')))
-
-    @property
-    def query_id(self) -> str:
-        warnings.warn("This field is deprecated. Use `id` instead.", DeprecationWarning)
-
-        return self.id