+++ rocketgram/api/shipping_query.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-import warnings
 from dataclasses import dataclass
 from typing import Optional
 
@@ -33,9 +32,3 @@ class ShippingQuery:
 
         return cls(data['id'], User.parse(data['from']), data['invoice_payload'],
                    ShippingAddress.parse(data['shipping_address']))
-
-    @property
-    def query_id(self) -> str:
-        warnings.warn("This field is deprecated. Use `id` instead.", DeprecationWarning)
-
-        return self.id