+++ rocketgram/api/update.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-import warnings
 from dataclasses import dataclass
 from typing import Dict, Optional
 
@@ -99,9 +98,3 @@ class Update:
         return cls(data, data['update_id'], update_type, message, edited_message, channel_post, edited_channel_post,
                    inline_query, chosen_inline_result, callback_query, shipping_query, pre_checkout_query, poll,
                    poll_answer, my_chat_member, chat_member, chat_join_request)
-
-    @property
-    def update_type(self) -> UpdateType:
-        warnings.warn("This field is deprecated. Use `type` instead.", DeprecationWarning)
-
-        return self.type