+++ rocketgram/api/video_chat_participants_invited.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-import warnings
 from dataclasses import dataclass
 from typing import Dict, Optional, List
 
@@ -25,9 +24,3 @@ class VideoChatParticipantsInvited:
             return None
 
         return cls([user.User.parse(u) for u in data['users']])
-
-
-def VoiceChatParticipantsInvited(*args, **kwargs):
-    warnings.warn("This class is deprecated. Use `VideoChatParticipantsInvited` instead.", DeprecationWarning)
-
-    return VideoChatParticipantsInvited(*args, **kwargs)